+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-01 13:07:34
+
+### Changed Files
+- `src/utils/policyCalculations.ts`
+
+### Statistics
+```
+ 1 file changed, 12 insertions(+), 5 deletions(-)
+```
+
 ## [Checkpoint] - 2026-01-01 12:56:08
 
 ### Changed Files